requirements.txt
atari-py
Box2D
gym
imageio
numpy
opencv-python
Pillow
pyglet
scipy
six
torch
tqdm
gym-retro
